RateMediaItem CM

+++ Program/ArtShare/ArtShare/Models/BookDetailsModel.cs
@@ -22,7 +22,7 @@ namespace ArtShare.Models
         public string Language { get; set; }
         public float? Price { get; set; }
         public int AccessRight { get; set; }
-        public float AvgRating { get; set; }
+        public double AvgRating { get; set; }
         public int RatingsGiven { get; set; }
     }
 }
