+++ Program/ArtShare/ArtShare/Models/IDetailsModel.cs
@@ -17,7 +17,7 @@ namespace ArtShare.Models
         List<string> Genres { get; set; }
         float? Price { get; set; }
         int AccessRight { get; set; }
-        float AvgRating { get; set; }
+        double AvgRating { get; set; }
         int RatingsGiven { get; set; }
 
     }
