+++ Program/ArtShare/ArtShare/Models/MusicDetailsModel.cs
@@ -21,7 +21,7 @@ namespace ArtShare.Models
         public string TrackLength { get; set; }
         public float? Price { get; set; }
         public int AccessRight { get; set; }
-        public float AvgRating { get; set; }
+        public double AvgRating { get; set; }
         public int RatingsGiven { get; set; }
     }
 }
